split the mod
